Reimplemented exception handling for this package.
For HTTP errors, a subclass of 'SlumberHttpBaseException' is thrown
depending on whether it's a client or server error.  The exception
instance is constructed with and provides 'response' and 'content'
attributes to expose the results of the HTTP request to the calling
code.

This is how httplib2 does things and how the original slumber does
things, but I modified the implementation of the exception classes to
make the attributes more explicit.

I also added 'SerializerEncodeError' and 'SerializerDecodeError' exception
classes to be raised when encoding the body of a request, or decoding the
content of a response.  'SerializerDecodeError' has a 'content' attribute
so calling code can see what couldn't be decoded.  'SerializerEncodeError'
doesn't do anything except pass through the error message from the
exception that was caught, but it made sense to me to include the class
to make things consistent.